feat: create components for each step
